backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
requests
python-dotenv
pydantic
tavily-python
ollama